Fix magic macros on MSVC, and fixed multithreaded rendering race condition

+++ src/dawn/renderer/VertexDecl.h
@@ -29,7 +29,7 @@ private:
                                 bool normalised);
     static void decodeAttributes(u16 encoded_attribute, Attribute& attribute, uint& count,
                                  AttributeType& type, bool& normalised);
-    static uint attributeTypeSize(AttributeType type);
+    static u16 attributeTypeSize(AttributeType type);
 
     // Attribute: 7
     // Count: 3
